pyproject.toml
[project]
name = "alpaca-mcp-server"
version = "0.1.0"
description = "MCP Server for Alpaca trading"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alpaca-py>=0.39.4",
    "mcp[cli]>=1.6.0",
    "pylint>=3.3.6",
    "python-dotenv>=1.1.0",
]
